1. Fixed in LookForAdditionalLocalExtremes function couple static_cast to correct data type (long -> int). 2. Removed some commented code.

+++ smarttools/NFmiLocalAreaMinMaxMask.cpp
@@ -869,8 +869,8 @@ static std::vector<LocalExtreme> LookForAdditionalLocalExtremes(
   ExtremeSearchCoreCounter searchCore;
 
   // Ei voi etsiä jos matriisi on liian pieni
-  if (static_cast<long>(subGridValues.NX()) > 3 * searchCoreRadiusInGridPoints &&
-      static_cast<long>(subGridValues.NY()) > 3 * searchCoreRadiusInGridPoints)
+  if (static_cast<int>(subGridValues.NX()) > 3 * searchCoreRadiusInGridPoints &&
+      static_cast<int>(subGridValues.NY()) > 3 * searchCoreRadiusInGridPoints)
   {
     // Käydään läpi kaikki ne pisteet mille voidaan tehdä laskut täydellä searchCore:lla
     for (size_t rowIndex = searchCoreRadiusInGridPoints;
@@ -1385,8 +1385,6 @@ std::vector<NFmiRect> NFmiLocalAreaMinMaxMask::CalculateLocalAreaCalculationBoun
     int subGridDecreaseIndexX,
     int subGridDecreaseIndexY)
 {
-  // auto gridSizeX = itsInfo->GridXNumber();
-  // auto gridSizeY = itsInfo->GridYNumber();
   std::vector<NFmiRect> boundaryVector;
   int gridPointIndexY = 0;
   for (int subGridYIndex = 0; subGridYIndex < subGridCountY; subGridYIndex++)
